requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langsmith>=0.1.0
orjson>=3.9.14,<4.0.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# NLP and entity extraction
spacy>=3.7.0
networkx>=3.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.7.0
